did the spawning and shooting logic for archot

+++ game_mode_draw.go
@@ -24,7 +24,7 @@ func (g *Game) GameModeDraw(screen *ebiten.Image) {
 
 	// draw player
 	playerOp := &ebiten.DrawImageOptions{}
-	playerOp.GeoM.Translate(float64(g.Player.X), float64(g.Player.Y))
+	playerOp.GeoM.Translate(g.Player.X, g.Player.Y)
 
 	screen.DrawImage(g.Player.GetSprite(), playerOp)
